requirements.txt
pydantic-ai>=0.0.10
openai
httpx>=0.25.0
beautifulsoup4
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0
eventregistry

